// src/app/@api/model/userRequestDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { OriginIdDto } from './originIdDto';
import { UserProfileDto } from './userProfileDto';


export interface UserRequestDto { 
    companyName?: string;
    email: string;
    firstName: string;
    jobTitle?: string;
    lastName: string;
    origin?: OriginIdDto;
    password: string;
    userProfile?: UserProfileDto;
    username: string;
}