// src/app/@api/model/userResponseDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { OriginIdDto } from './originIdDto';


export interface UserResponseDto { 
    email?: string;
    firstName?: string;
    jobTitle?: string;
    lastName?: string;
    origin?: OriginIdDto;
    status?: number;
}